Atualizado. Falta ajuste inserts

+++ API_Juntos.Application/Models/Pedidos/ListarPedidoPorId/ListarPedidoporIdProdutosDetalhadosResponse.cs
@@ -6,7 +6,7 @@
         public string NomeProduto { get; set; }
         public decimal ValorUnitario { get; set; }
         public int Quantidade { get; set; }
-        public int totalParcial { get; set; }
+        public decimal totalParcial { get; set; }
 
     }
 }
